refactor(assets): merge generate-sdk.mjs into ensure-sdk.mjs behind --check-spec
One file per repo instead of two (plus the sdk-lib.mjs split): ensure-sdk.mjs
now exports its functions and only runs as a CLI when executed directly, so
generate:sdk becomes `ensure-sdk.mjs --check-spec` and EE's wrapper imports
the same functions unchanged.
