+++ ui/package.json
@@ -16,7 +16,7 @@
         "build:topology": "npm run build --workspace=@kestra-io/topology",
         "build:sdk": "npm run build --workspace=@kestra-io/hey-api-plugin && npm run build --workspace=@kestra-io/kestra-sdk",
         "generate:openapi-spec": "cd .. && ./gradlew :webserver:generateOpenapiSpec",
-        "generate:sdk": "node scripts/generate-sdk.mjs",
+        "generate:sdk": "node scripts/ensure-sdk.mjs --check-spec",
         "preview": "vite preview",
         "precheck:types": "node scripts/ensure-sdk.mjs",
         "check:types": "vue-tsc --build packages/design-system/tsconfig.app.json && echo \"design-system compiled\" \\\n && vue-tsc --build tsconfig.app.json && echo \"app checked\" \\\n && vue-tsc --project tsconfig.test.json && echo \"test checked\"",
